requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
jinja2==3.1.3
python-multipart==0.0.6
google-generativeai==0.3.2
pydantic==2.6.1
PyPDF2==3.0.1
python-dotenv==1.0.0
starlette==0.36.1
PyMuPDF==1.23.8
pdf2image==1.16.3
pytesseract==0.3.10
python-magic==0.4.27 